Exercise3/requirements.txt
moderngl
numpy
pyqt5
Pillow
pyopengl
